requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==5.2.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.6
Django==4.1.5
django-allauth==0.52.0
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-redis==5.2.0
django-sesame==3.1
django-sslserver==0.22
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
drf-yasg==1.21.4
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
jsonpickle==3.0.1
kombu==5.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
python3-openid==3.2.0
pytz==2022.7
redis==4.4.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
